add acceptance test helper
